Moved constant translation to clause translator.
